+++ Hbuilder_Example/nativeplugins/GCUniPlugin/ios/FTMobileSDK.xcframework/ios-arm64/FTMobileSDK.framework/Headers/FTResourceMetricsModel.h
@@ -26,12 +26,16 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, strong) NSNumber *resource_first_byte;
 ///资源加载时间 duration(responseEnd-fetchStartDate)
 @property (nonatomic, strong) NSNumber *duration;
+///响应结果大小 response data size
+@property (nonatomic, strong, nullable) NSNumber *responseSize;
+///远程地址
+@property (nonatomic, copy) NSString *remoteAddress;
 /// 初始化方法
 ///
 /// - Parameters:
 ///   - metrics: SessionTaskMetric
 /// - Returns: metrics 实例.
--(instancetype)initWithTaskMetrics:(NSURLSessionTaskMetrics *)metrics API_AVAILABLE(ios(10.0),macosx(10.12));
+-(instancetype)initWithTaskMetrics:(NSURLSessionTaskMetrics *)metrics API_AVAILABLE(ios(10.0),macos(10.12));
 
 @end