+++ Hbuilder_Example/nativeplugins/GCUniPlugin/ios/FTMobileSDK.xcframework/ios-arm64_x86_64-simulator/FTMobileSDK.framework/Headers/FTMobileConfig.h
@@ -52,7 +52,7 @@ typedef NS_ENUM(NSUInteger, FTMonitorFrequency) {
     FTMonitorFrequencyRare,
 };
 /// 网络链路追踪使用类型
-typedef NS_ENUM(NSInteger, FTNetworkTraceType) {
+typedef NS_ENUM(NSUInteger, FTNetworkTraceType) {
     /// datadog trace
     FTNetworkTraceTypeDDtrace,
     /// zipkin multi header
@@ -79,6 +79,15 @@ typedef NS_ENUM(NSInteger, FTEnv) {
     /// 本地环境
     FTEnvLocal,
 };
+/// 数据同步大小
+typedef NS_ENUM(NSUInteger, FTSyncPageSize) {
+    /// MINI 5
+    FTSyncPageSizeMini = 0,
+    /// MEDIUM 10
+    FTSyncPageSizeMedium,
+    /// MAX 50
+    FTSyncPageSizeMax,
+};
 /// 日志废弃策略
 typedef NS_ENUM(NSInteger, FTLogCacheDiscard)  {
     /// 默认，当日志数据数量大于最大值（5000）时，新数据不进行写入
@@ -86,8 +95,23 @@ typedef NS_ENUM(NSInteger, FTLogCacheDiscard)  {
     /// 当日志数据大于最大值时,废弃旧数据
     FTDiscardOldest
 };
-
+/// RUM废弃策略
+typedef NS_ENUM(NSInteger, FTRUMCacheDiscard)  {
+    /// 默认，当日志数据数量大于最大值（100_000）时，新数据不进行写入
+    FTRUMDiscard,
+    /// 当日志数据大于最大值时,废弃旧数据
+    FTRUMDiscardOldest
+};
+/// DB废弃策略
+typedef NS_ENUM(NSInteger, FTDBCacheDiscard)  {
+    /// 默认，当数据库存储大于最大值(默认100MB)时，新数据不进行写入
+    FTDBDiscard,
+    /// 当数据库存储大于最大值,废弃旧数据
+    FTDBDiscardOldest
+};
 NS_ASSUME_NONNULL_BEGIN
+/// RUM 过滤 resource 回调，返回：NO 表示要采集，YES 表示不需要采集。
+typedef BOOL(^FTResourceUrlHandler)(NSURL * url);
 
 /// logger 功能配置项
 @interface FTLoggerConfig : NSObject
@@ -103,6 +127,8 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) BOOL enableCustomLog;
 /// 是否将自定义日志在控制台打印
 @property (nonatomic, assign) BOOL printCustomLogToConsole; 
+/// 日志最大缓存量, 最低设置为 1000  默认 5000
+@property (nonatomic, assign) int logCacheLimitCount;
 /// 采集自定义日志的状态数组，默认为全采集
 ///
 /// 例: @[@(FTStatusInfo),@(FTStatusError)]
@@ -135,10 +161,17 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) BOOL enableTraceUserView;
 /// 设置是否追踪用户网络请求  (仅作用于native http)
 @property (nonatomic, assign) BOOL enableTraceUserResource;
+/// 设置是否采集网络请求 Host IP (仅作用于native http，iOS 13及以上)
+@property (nonatomic, assign) BOOL enableResourceHostIP;
+/// 自定义采集 resource 规则。
+/// 根据请求资源 url 判断是否需要采集对应资源数据，默认都采集。 返回：NO 表示要采集，YES 表示不需要采集。
+@property (nonatomic, copy) FTResourceUrlHandler resourceUrlHandler;
 /// 设置是否需要采集崩溃日志
 @property (nonatomic, assign) BOOL enableTrackAppCrash;
 /// 设置是否需要采集卡顿
 @property (nonatomic, assign) BOOL enableTrackAppFreeze;
+/// 设置卡顿的阈值。单位毫秒 100 < freezeDurationMs ，默认 250ms
+@property (nonatomic, assign) long freezeDurationMs;
 /// 设置是否需要采集 ANR
 ///
 /// runloop 采集主线程卡顿
@@ -153,6 +186,15 @@ NS_ASSUME_NONNULL_BEGIN
 ///
 /// 保留标签:特殊 key - track_id (用于追踪功能)
 @property (nonatomic, strong) NSDictionary<NSString*,NSString*> *globalContext;
+/// RUM 最大缓存量,  默认 100_000
+@property (nonatomic, assign) int rumCacheLimitCount;
+/// RUM废弃策略
+@property (nonatomic, assign) FTRUMCacheDiscard rumDiscardType;
+
+/// 开启采集卡顿并设置卡顿的阈值。
+/// - Parameter enableTrackAppFreeze: 设置是否需要采集卡顿
+/// - Parameter freezeDurationMs: 卡顿的阈值，单位毫秒 100 < freezeDurationMs ，默认 250ms
+-(void)setEnableTrackAppFreeze:(BOOL)enableTrackAppFreeze freezeDurationMs:(long)freezeDurationMs;
 @end
 /// Trace 功能配置项
 @interface FTTraceConfig : NSObject
@@ -174,7 +216,16 @@ NS_ASSUME_NONNULL_BEGIN
 @interface FTMobileConfig : NSObject
 /// 指定初始化方法，设置 metricsUrl
 /// - Parameter metricsUrl: 数据上报地址
-- (instancetype)initWithMetricsUrl:(nonnull NSString *)metricsUrl;
+- (instancetype)initWithMetricsUrl:(NSString *)metricsUrl DEPRECATED_MSG_ATTRIBUTE("已过时，请使用 -initWithDatakitUrl: 替换");
+
+/// 本地环境部署，设置 datakitUrl
+/// - Parameter datakitUrl: datakit 数据上报地址
+- (instancetype)initWithDatakitUrl:(NSString *)datakitUrl;
+
+/// 使用公网 DataWay 部署，设置 datawayUrl 与 clientToken
+/// - Parameter datawayUrl: datawayUrl 数据上报地址
+/// - Parameter clientToken: dataway token
+- (instancetype)initWithDatawayUrl:(NSString *)datawayUrl clientToken:(NSString *)clientToken;
 
 /// 禁用 init 初始化
 - (instancetype)init NS_UNAVAILABLE;
@@ -182,15 +233,38 @@ NS_ASSUME_NONNULL_BEGIN
 /// 禁用 new 初始化
 + (instancetype)new NS_UNAVAILABLE;
 /// 数据上报地址
-@property (nonatomic, copy) NSString *metricsUrl;
+@property (nonatomic, copy) NSString *metricsUrl DEPRECATED_MSG_ATTRIBUTE("已过时，请使用 datakitUrl 替换");
+/// 数据上报 datakit 地址
+@property (nonatomic, copy) NSString *datakitUrl;
+/// 数据上报 dataway 地址
+@property (nonatomic, copy) NSString *datawayUrl;
+/// client token
+@property (nonatomic, copy) NSString *clientToken;
 /// 设置自定义环境字段。
 @property (nonatomic, copy) NSString *env;
 /// 设置是否允许 SDK 打印 Debug 日志。
 @property (nonatomic, assign) BOOL enableSDKDebugLog;
-/// 应用版本号。
-@property (nonatomic, copy) NSString *version;
+/// 应用版本号。默认`CFBundleShortVersionString`值
+@property (nonatomic, copy) NSString *version DEPRECATED_MSG_ATTRIBUTE("已废弃，version 将统一使用`CFBundleShortVersionString`值");
 /// 所属业务或服务的名称 默认：df_rum_ios
 @property (nonatomic, copy) NSString *service;
+/// 数据是否进行自动同步上传 默认：YES
+@property (nonatomic, assign) BOOL autoSync;
+/// 数据同步时每条请求同步条数,最小值 5 默认：10
+@property (nonatomic, assign) int syncPageSize;
+/// 数据同步时每条请求间隔时间 单位毫秒 0< syncSleepTime <5000
+@property (nonatomic, assign) int syncSleepTime;
+/// 数据同步时是否开启数据整数兼容
+@property (nonatomic, assign) BOOL enableDataIntegerCompatible;
+/// 设置内部数据同步时是否开启压缩 默认: NO
+@property (nonatomic, assign) BOOL compressIntakeRequests;
+/// 开启使用 db 限制数据大小
+@property (nonatomic, assign) BOOL enableLimitWithDbSize;
+/// db 缓存限制大小,默认 100MB,单位 byte
+@property (nonatomic, assign) long dbCacheLimit;
+/// 数据库废弃策略
+@property (nonatomic, assign) FTDBCacheDiscard dbDiscardType;
+
 /// 设置 SDK 全局 tag
 ///
 /// 保留标签： sdk_package_flutter、sdk_package_react_native
@@ -202,6 +276,10 @@ NS_ASSUME_NONNULL_BEGIN
 /// 根据提供的 FTEnv 类型设置 env
 /// - Parameter envType: 环境
 - (void)setEnvWithType:(FTEnv)envType;
+/// 根据提供的 FTSyncPageSize 类型设置 syncPageSize
+/// - Parameter pageSize: 数据同步大小
+- (void)setSyncPageSizeWithType:(FTSyncPageSize)pageSize;
+
 @end
 
 NS_ASSUME_NONNULL_END